requirements.txt
# Flask Framework
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# Production WSGI Server
gunicorn==21.2.0

# Firebase & Google Cloud
firebase-admin==6.4.0
google-cloud-firestore==2.13.1
google-cloud-core==2.4.1
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Google AI (Gemini)
google-generativeai==0.3.2
google-ai-generativelanguage==0.4.0

# YouTube Data API
google-api-python-client==2.108.0

# MFA & TOTP
pyotp==2.9.0
qrcode[pil]==7.4.2

# HTTP Requests & API Integration
requests==2.31.0
urllib3==2.1.0

# Environment & Configuration
python-dotenv==1.0.0

# Data Validation & Processing
email-validator==2.1.0
validators==0.22.0
marshmallow==3.20.2

# Date & Time Handling
python-dateutil==2.8.2

# Logging & Monitoring
structlog==23.2.0

# Security & Cryptography
cryptography>=41.0.0
PyJWT==2.8.0

# Data Processing & Utilities
pandas==2.1.4
numpy==1.26.2

# Email & SMTP
secure-smtplib==0.1.1

# Development & Testing (optional)
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Code Quality (optional)
black==23.12.0
flake8==6.1.0
isort==5.13.2

# Performance & Caching
redis==5.0.1
cachetools==5.3.2

# Additional Utilities
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3
Jinja2==3.1.2
blinker==1.7.0